[mysql] Separate binlog-connector implements
