+++ spinaltap-mysql/src/main/java/com/airbnb/spinaltap/mysql/MysqlDestinationMetrics.java
@@ -7,6 +7,7 @@ package com.airbnb.spinaltap.mysql;
 import com.airbnb.common.metrics.TaggedMetricRegistry;
 import com.airbnb.spinaltap.Mutation;
 import com.airbnb.spinaltap.common.destination.DestinationMetrics;
+import com.airbnb.spinaltap.mysql.binlog_connector.BinaryLogConnectorSource;
 import com.airbnb.spinaltap.mysql.mutation.MysqlMutationMetadata;
 import com.google.common.base.Preconditions;
 import java.util.HashMap;
@@ -16,7 +17,7 @@ import lombok.NonNull;
 /**
  * Responsible for metrics collection on operations of the {@link
  * com.airbnb.spinaltap.common.destination.Destination} and associated components for a given {@link
- * MysqlSource}.
+ * BinaryLogConnectorSource}.
  */
 public final class MysqlDestinationMetrics extends DestinationMetrics {
   private static final String DATABASE_NAME_TAG = "database_name";
