+++ spinaltap-mysql/src/main/java/com/airbnb/spinaltap/mysql/config/AbstractMysqlConfiguration.java
@@ -6,18 +6,23 @@ package com.airbnb.spinaltap.mysql.config;
 
 import com.airbnb.spinaltap.common.config.DestinationConfiguration;
 import com.airbnb.spinaltap.common.config.SourceConfiguration;
+import com.airbnb.spinaltap.mysql.MysqlSource;
+
 import java.util.List;
 
+/**
+ * Represents the base configuration for a {@link MysqlSource}.
+ */
 public abstract class AbstractMysqlConfiguration extends SourceConfiguration {
   public AbstractMysqlConfiguration(
-      String name,
-      String type,
-      String instanceTag,
-      DestinationConfiguration destinationConfiguration) {
+      final String name,
+      final String type,
+      final String instanceTag,
+      final DestinationConfiguration destinationConfiguration) {
     super(name, type, instanceTag, destinationConfiguration);
   }
 
-  public AbstractMysqlConfiguration(String type, String instanceTag) {
+  public AbstractMysqlConfiguration(final String type,final String instanceTag) {
     super(type, instanceTag);
   }
