+++ spinaltap-mysql/src/main/java/com/airbnb/spinaltap/mysql/config/MysqlSchemaStoreConfiguration.java
@@ -4,12 +4,17 @@
  */
 package com.airbnb.spinaltap.mysql.config;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
-import lombok.Data;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * Represents configuration for a {@link com.airbnb.spinaltap.mysql.schema.MysqlSchemaStore}.
+ */
 @Data
 public class MysqlSchemaStoreConfiguration {
   @NotNull @JsonProperty private String host;
